Create hello_world.b
Here's a "Hello, World!" program in B. B is a simple programming language that was developed by Ken Thompson in the early 1970s.

When you run this B program, it will print "Hello, World!" to the console. Please note that B is a historical programming language and is not commonly used today.